Version 2.0 - First try. The old API was deprecated and internally replaced with equivalents using the new event based API in order to hopefully stay compatible to plugins still using the old API. The next update will remove those deprecated methods..
IndividualSigns does now simply call a SignSendEvent for each player
being about to receiving a sign update packet. This event can be
cancelled and gives your more freedom on what can be done compared to
before (your are not limited anymore on simple key->value replacements,
but you can for example search for dynamic keys or replace the text of
the whole sign, etc).

However, you will now have to process the lines yourself (checking for
the key and replacing found matches), and you have to check the
permission for the sign creation yourself as well.
OR you use the included SimpleChanger class, which basicly provides the
same features like the old, now deprecated, Changer class: you simple
create a new SimpleChanger and fill in the getValue() method (which now
has a new 'line' parameter), and the class will do permission checks and
simple key-value replacement for you, just like the old API, but more
nicely.