Fixed: The sign text placeholders were not getting properly replaced in chunk packets.
With the 1.18 changes to how we handle the chunk packet, we need to apply the final outgoing tile entity info list to the outgoing packet after all modifications have been applied. Due to the use of the tile entity info converter, this list is no longer a view on the actual list that is stored by the packet, but is instead a separate list that is applied indirectly to the outgoing packet via a conversion.